refactor STM.internal to separate TID and RefLog implementations
